requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
certifi==2024.12.14
cfgv==3.4.0
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
dependency-injector==4.45.0
distlib==0.3.9
Django==5.1.5
django-rest-swagger==2.2.0
djangorestframework==3.15.2
dnspython==2.7.0
drf-yasg==1.21.8
filelock==3.16.1
gunicorn==23.0.0
identify==2.6.5
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mongoengine==0.29.1
nodeenv==1.9.1
openapi-codec==1.3.2
packaging==24.2
platformdirs==4.3.6
pre_commit==4.0.1
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pymongo==4.10.1
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
simplejson==3.19.3
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
virtualenv==20.29.1
